+++ hive-io-exp-core/src/test/java/com/facebook/hiveio/output/OutputTest.java
@@ -124,7 +124,7 @@ public class OutputTest extends HiveIOTestBase {
     assertTrue(readRecords.hasNext());
 
     HiveReadableRecord record = readRecords.next();
-    assertEquals(Long.class, record.get(0).getClass());
+    assertEquals(Integer.class, record.get(0).getClass());
     assertEquals(Double.class, record.get(1).getClass());
     assertEquals(1, record.getInt(0));
     assertEquals(1.1, record.getDouble(1));
